Rename linear module to reverse_headers
